added template for testing
