+++ jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+  preset: 'react-native',
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+  transformIgnorePatterns: [
+    'node_modules/(?!(jest-)?@?react-native' +
+      '|react-navigation-tabs' +
+      '|react-native-splash-screen' +
+      '|react-native-screens' +
+      '|react-native-reanimated' +
+      '|react-native-card-flip' +
+      '|@react-navigation' +
+      '|@react-native-community' +
+      '/)'
+  ]
+};
