+++ package.json
@@ -15,6 +15,7 @@
     "@react-navigation/bottom-tabs": "^5.11.10",
     "@react-navigation/native": "^5.9.4",
     "@react-navigation/stack": "^5.14.4",
+    "@testing-library/react-native": "^7.2.0",
     "@ui-kitten/components": "^5.1.0",
     "expo": "~41.0.1",
     "expo-google-app-auth": "^8.1.7",
